+++ packages/core/src/editor/index.ts
@@ -1,7 +1,9 @@
-export type { EditorSharedState, EditorSharedStateKey } from './shared-state'
-export { EDITOR_SHARED_STATE_KEYS } from './shared-state'
-export type { EditorViewState, EditorViewStateKey } from './view-state'
-export { copyEditorViewState, EDITOR_VIEW_STATE_KEYS, pickEditorViewState } from './view-state'
+export { createDefaultEditorSharedState } from './state/shared'
+export {
+  copyEditorViewState,
+  createDefaultEditorViewState,
+  pickEditorViewState
+} from './state/view'
 export { createDefaultEditorState, createEditor } from './create'
 export type { Editor } from './create'
 export { createTextActions } from './text'
@@ -16,6 +18,8 @@ export type {
   EditorEvents,
   EditorOptions,
   EditorState,
+  EditorSharedState,
+  EditorViewState,
   FigmaClipboardImageResolver,
   Tool
 } from './types'
