+++ packages/core/src/editor/state/index.ts
@@ -1,0 +1,11 @@
+import type { EditorState } from '#core/editor/types'
+
+import { createDefaultEditorSharedState } from './shared'
+import { createDefaultEditorViewState } from './view'
+
+export function createDefaultEditorState(pageId: string): EditorState {
+  return {
+    ...createDefaultEditorSharedState(),
+    ...createDefaultEditorViewState(pageId)
+  }
+}
